hoemPage

+++ lib/core/constants/appSizes.dart
@@ -1,6 +1,6 @@
 class appsizes {
   static double borderRadius = 100.0;
-  static int padding = 16;
+  static double padding = 16;
   static int spaceSmall = 8;
   static int spaceMedium = 16;
